Fix backport of table_exists? in 14accdf8d1
The backport of table_exists? expects tables to accept an arity of 2,
whereas the version in the Mysql2Adapter only has an arity of 1. The
second parameter provides for querying tables in databases other than
the currently selected database.

This commit fixes this by backporting the tables method from the
MysqlAdapter in Rails 3.0.x.